+++ CHANGELOG.md
@@ -1,5 +1,48 @@
 # Change Log 
 
+## [1.17.7](https://github.com/alibaba-fusion/next/compare/1.17.6...1.17.7) (2019-08-21)
+
+
+### Bug Fixes
+
+* **Checkbox:** checkbox will update indeterminate state when in group ([8b9582a](https://github.com/alibaba-fusion/next/commit/8b9582a))
+* **Overlay:** remove position absolute. fix [#1016](https://github.com/alibaba-fusion/next/issues/1016) ([7d87e86](https://github.com/alibaba-fusion/next/commit/7d87e86))
+* **Select:** fix onSearchClear function ([12f2a92](https://github.com/alibaba-fusion/next/commit/12f2a92))
+* **Select:** readonly ([0d732d6](https://github.com/alibaba-fusion/next/commit/0d732d6))
+* **Tree:** undined of parent ([f38f440](https://github.com/alibaba-fusion/next/commit/f38f440))
+
+
+
+
+## [1.17.6](https://github.com/alibaba-fusion/next/compare/1.17.5...1.17.6) (2019-08-15)
+
+
+### Bug Fixes
+
+* **Menu:** value in selectedKeys doesn't exit in k2n ([fef80cd](https://github.com/alibaba-fusion/next/commit/fef80cd))
+
+
+
+
+## [1.17.5](https://github.com/alibaba-fusion/next/compare/1.17.4...1.17.5) (2019-08-14)
+
+
+### Bug Fixes
+
+* **Field:** bug of resetToDefault withtou initValue. close [#1001](https://github.com/alibaba-fusion/next/issues/1001) ([a76b05c](https://github.com/alibaba-fusion/next/commit/a76b05c))
+* **Input:** support null to reset value ([0ecc5c2](https://github.com/alibaba-fusion/next/commit/0ecc5c2))
+* **Pagination:** react warning use setTimout event ([0e92ade](https://github.com/alibaba-fusion/next/commit/0e92ade))
+
+
+### Features
+
+* **Animate:** add slide animations ([407dc5b](https://github.com/alibaba-fusion/next/commit/407dc5b))
+* **Drawer:** make drawer configurable ([fbc4ad2](https://github.com/alibaba-fusion/next/commit/fbc4ad2))
+* **Locale:** add Drawer words ([1be3ea8](https://github.com/alibaba-fusion/next/commit/1be3ea8))
+
+
+
+
 ## [1.17.4](https://github.com/alibaba-fusion/next/compare/1.17.3...1.17.4) (2019-08-13)